Initial commit for autonomy integration
