+++ src/robot2017/src/teleop_exec.h
@@ -3,20 +3,26 @@
 //calls appropriate motor functions
 
 #include "robot_msgs/Teleop.h"
+#include "robot_msgs/Autonomy.h"
 #include "motors.h"
 
-#ifndef TELEOP_EXEC_H
-#define TELEOP_EXEC_H
+#ifndef ROBOT_EXEC_H
+#define ROBOT_EXEC_H
 
-class TeleopExec
+class RobotExec
 {
     private:
         bool dead;
+        bool autonomyActive;
         float leftRatio;
         float rightRatio;
 
     public:
+        void teleopReceived(const robot_msgs::Teleop& cmd);
+        void autonomyReceived(const robot_msgs::Autonomy& cmd);
+
         void teleopExec(const robot_msgs::Teleop& cmd);
+        void autonomyExec(const robot_msgs::Autonomy& cmd);
         void killMotors();
 };
